CrewAI/CrewAI-YT_Video/requirements.txt
crewai
crewai[tools]
python-dotenv
youtube-transcript-api
yt_dlp
